+++ Nanagong/Nanagong/CreateUserSessionManager/KakaoSessionManager.swift
@@ -9,6 +9,7 @@ import Combine
 import Foundation
 import KakaoSDKAuth
 import KakaoSDKCommon
+import KakaoSDKUser
 
 enum KakaoSessionManagerError: Error {
     
@@ -18,7 +19,7 @@ enum KakaoSessionManagerError: Error {
 
 final class KakaoSessionManager {
     
-    private let combineWrapper: KakaoSDKCombineWrapper = .init()
+    private let loginResultPublisher: PassthroughSubject<OAuthToken, KakaoSessionManagerError> = .init()
     
     func initSDK() {
         let appID = Bundle.main.object(forInfoDictionaryKey: "KAKAO_APPID") as? String ?? ""
@@ -33,7 +34,36 @@ final class KakaoSessionManager {
         return AuthController.handleOpenUrl(url: url)
     }
     
-    func loginWithKakaoTalk() -> AnyPublisher<OAuthToken, KakaoSessionManagerError> {
-        return combineWrapper.loginWithKakao()
+    func loginWithKakao() -> AnyPublisher<OAuthToken, KakaoSessionManagerError> {
+        if UserApi.isKakaoTalkLoginAvailable() {
+            loginWithKakaoTalk()
+        } else {
+            loginWithKakaoAccount()
+        }
+        
+        return loginResultPublisher.eraseToAnyPublisher()
+    }
+    
+    private func loginWithKakaoTalk() {
+        UserApi.shared.loginWithKakaoTalk { [weak self] token, error in
+            self?.processLoginResult(with: (token, error))
+        }
+    }
+    
+    private func loginWithKakaoAccount() {
+        UserApi.shared.loginWithKakaoAccount {[weak self] token, error in
+            self?.processLoginResult(with: (token, error))
+        }
+    }
+    
+    private func processLoginResult(with result: (token: OAuthToken?, error: Error?)) {
+        if let error = result.error {
+            loginResultPublisher.send(completion: .failure(.kakaoError(error)))
+        } else if let token = result.token {
+            loginResultPublisher.send(token)
+            loginResultPublisher.send(completion: .finished)
+        } else {
+            loginResultPublisher.send(completion: .failure(.unknownError))
+        }
     }
 }
